requirements.txt
streamlit
finnhub-python
pandas
numpy
scikit-learn
tensorflow
matplotlib
plotly
textblob
python-dotenv
sqlite3-binary
